% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MdgmModel}
\alias{MdgmModel}
\title{MDGM Model}
\description{
MDGM Model

MDGM Model
}
\details{
R6 class wrapping a C++ MDGM model. Use \code{\link[=mdgm_model]{mdgm_model()}} to create instances.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MdgmModel-new}{\code{MdgmModel$new()}}
\item \href{#method-MdgmModel-has_emission}{\code{MdgmModel$has_emission()}}
\item \href{#method-MdgmModel-nvertices}{\code{MdgmModel$nvertices()}}
\item \href{#method-MdgmModel-ncolors}{\code{MdgmModel$ncolors()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MdgmModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MdgmModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new MdgmModel. Use \code{\link[=mdgm_model]{mdgm_model()}} instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$new(ptr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ptr}}{External pointer to a C++ Model object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MdgmModel-has_emission"></a>}}
\if{latex}{\out{\hypertarget{method-MdgmModel-has_emission}{}}}
\subsection{Method \code{has_emission()}}{
Check if the model has an emission layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$has_emission()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MdgmModel-nvertices"></a>}}
\if{latex}{\out{\hypertarget{method-MdgmModel-nvertices}{}}}
\subsection{Method \code{nvertices()}}{
Get the number of vertices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$nvertices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MdgmModel-ncolors"></a>}}
\if{latex}{\out{\hypertarget{method-MdgmModel-ncolors}{}}}
\subsection{Method \code{ncolors()}}{
Get the number of colors (categories).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$ncolors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer.
}
}
}
