% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mrf}
\alias{mrf}
\title{MRF spatial configuration}
\usage{
mrf(method = c("exchange", "pseudo_likelihood"), n_aux_sweeps = 200L)
}
\arguments{
\item{method}{Inference method for psi: \code{"exchange"} (exact, using the
exchange algorithm) or \code{"pseudo_likelihood"} (approximate).}

\item{n_aux_sweeps}{Number of auxiliary Gibbs sweeps for the exchange
algorithm (default 200). Ignored for pseudo-likelihood.}
}
\value{
A list of class \code{mrf_config}.
}
\description{
Creates a configuration object for a Markov random field (Potts model)
spatial field. Used as the \code{spatial} argument to \code{\link[=srf_model]{srf_model()}}.
}
\examples{
mrf(method = "exchange")
mrf(method = "pseudo_likelihood")
}
