% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_undirected_graph.R
\name{nug_from_adj_list}
\alias{nug_from_adj_list}
\title{Create an undirected graph from an adjacency list}
\usage{
nug_from_adj_list(adj_list, ...)
}
\arguments{
\item{adj_list}{A list of integer vectors. The i-th element contains the
1-indexed neighbors of vertex i.}

\item{...}{Additional arguments: \code{seed} (integer) for reproducible RNG.}
}
\value{
A \code{NaturalUndirectedGraph} object.
}
\description{
Create an undirected graph from an adjacency list
}
\examples{
# Path graph: 1 -- 2 -- 3
adj <- list(c(2L), c(1L, 3L), c(2L))
g <- nug_from_adj_list(adj)
g$nvertices()
}
\seealso{
Other graph-construction: 
\code{\link{nug_from_adj_mat}()},
\code{\link{nug_from_edge_list}()}
}
\concept{graph-construction}
