% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_undirected_graph.R
\name{nug_from_adj_mat}
\alias{nug_from_adj_mat}
\title{Create an undirected graph from an adjacency matrix}
\usage{
nug_from_adj_mat(adj_mat, ...)
}
\arguments{
\item{adj_mat}{A square symmetric matrix. Nonzero entries indicate edges;
their values are used as edge weights. Self-loops are not allowed.}

\item{...}{Additional arguments: \code{seed} (integer) for reproducible RNG.}
}
\value{
A \code{NaturalUndirectedGraph} object.
}
\description{
Create an undirected graph from an adjacency matrix
}
\examples{
# 4-vertex path graph from adjacency matrix
A <- matrix(0, 4, 4)
A[1, 2] <- A[2, 1] <- 1
A[2, 3] <- A[3, 2] <- 1
A[3, 4] <- A[4, 3] <- 1
g <- nug_from_adj_mat(A)
g$nedges()
}
\seealso{
Other graph-construction: 
\code{\link{nug_from_adj_list}()},
\code{\link{nug_from_edge_list}()}
}
\concept{graph-construction}
