% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{SrfModel}
\alias{SrfModel}
\alias{MdgmModel}
\title{Spatial Random Field Model}
\description{
Spatial Random Field Model

Spatial Random Field Model
}
\details{
R6 class wrapping a C++ spatial random field model. Use \code{\link[=srf_model]{srf_model()}} to
create instances.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SrfModel-new}{\code{SrfModel$new()}}
\item \href{#method-SrfModel-has_emission}{\code{SrfModel$has_emission()}}
\item \href{#method-SrfModel-nvertices}{\code{SrfModel$nvertices()}}
\item \href{#method-SrfModel-ncolors}{\code{SrfModel$ncolors()}}
\item \href{#method-SrfModel-emission_type}{\code{SrfModel$emission_type()}}
\item \href{#method-SrfModel-model_type}{\code{SrfModel$model_type()}}
\item \href{#method-SrfModel-get_ptr}{\code{SrfModel$get_ptr()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new SrfModel. Use \code{\link[=srf_model]{srf_model()}} instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SrfModel$new(ptr, model_type = "mdgm")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ptr}}{External pointer to a C++ Model object.}

\item{\code{model_type}}{Character string: \code{"mdgm"} or \code{"mrf"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-has_emission"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-has_emission}{}}}
\subsection{Method \code{has_emission()}}{
Check if the model has an emission layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SrfModel$has_emission()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-nvertices"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-nvertices}{}}}
\subsection{Method \code{nvertices()}}{
Get the number of vertices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SrfModel$nvertices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-ncolors"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-ncolors}{}}}
\subsection{Method \code{ncolors()}}{
Get the number of colors (categories).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SrfModel$ncolors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-emission_type"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-emission_type}{}}}
\subsection{Method \code{emission_type()}}{
Get the emission family type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SrfModel$emission_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string (\code{"bernoulli"}, \code{"gaussian"}, \code{"poisson"}) or
\code{NULL} for standalone models.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-model_type"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-model_type}{}}}
\subsection{Method \code{model_type()}}{
Get the spatial model type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SrfModel$model_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string: \code{"mdgm"} or \code{"mrf"}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-get_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-get_ptr}{}}}
\subsection{Method \code{get_ptr()}}{
Get the internal C++ pointer. For internal use only.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SrfModel$get_ptr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
External pointer.
}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SrfModel-new}{\code{MdgmModel$new()}}
\item \href{#method-SrfModel-has_emission}{\code{MdgmModel$has_emission()}}
\item \href{#method-SrfModel-nvertices}{\code{MdgmModel$nvertices()}}
\item \href{#method-SrfModel-ncolors}{\code{MdgmModel$ncolors()}}
\item \href{#method-SrfModel-emission_type}{\code{MdgmModel$emission_type()}}
\item \href{#method-SrfModel-model_type}{\code{MdgmModel$model_type()}}
\item \href{#method-SrfModel-get_ptr}{\code{MdgmModel$get_ptr()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$new(ptr, model_type = "mdgm")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-has_emission"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-has_emission}{}}}
\subsection{Method \code{has_emission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$has_emission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-nvertices"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-nvertices}{}}}
\subsection{Method \code{nvertices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$nvertices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-ncolors"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-ncolors}{}}}
\subsection{Method \code{ncolors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$ncolors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-emission_type"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-emission_type}{}}}
\subsection{Method \code{emission_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$emission_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-model_type"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-model_type}{}}}
\subsection{Method \code{model_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$model_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SrfModel-get_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-SrfModel-get_ptr}{}}}
\subsection{Method \code{get_ptr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MdgmModel$get_ptr()}\if{html}{\out{</div>}}
}

}
}
