% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undirected_graph.R
\name{ug_from_adj_list}
\alias{ug_from_adj_list}
\title{Creates an undirected graph from an adjacency list}
\usage{
ug_from_adj_list(adj_list, ...)
}
\arguments{
\item{adj_list}{Adjacency list as a list of integer vectors. The i-th element of the list
contains the neighbors of vertex i.}

\item{...}{Additional arguments passed to the \code{UndirectedGraph} constructor for setting up the
C++ mdgm::RNG. Can set the \code{seed} as an integer.}
}
\value{
An \code{UndirectedGraph} object.
}
\description{
Creates an undirected graph from an adjacency list
}
