% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undirected_graph.R
\name{ug_from_adj_mat}
\alias{ug_from_adj_mat}
\title{Creates an undirected graph from an adjacency matrix}
\usage{
ug_from_adj_mat(adj_mat, ...)
}
\arguments{
\item{adj_mat}{Adjacency matrix as a square symmetric matrix. Nonzero entries indicate the edges
and edge weights. Self-loops are not allowed.}

\item{...}{Additional arguments passed to the \code{UndirectedGraph} constructor for setting up the
C++ mdgm::RNG. Can set the \code{seed} as an integer.}
}
\value{
An \code{UndirectedGraph} object.
}
\description{
Creates an undirected graph from an adjacency matrix
}
