% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undirected_graph.R
\name{ug_from_edge_list}
\alias{ug_from_edge_list}
\title{Creates an undirected graph from an edge list}
\usage{
ug_from_edge_list(n, edges, ...)
}
\arguments{
\item{n}{Number of vertices}

\item{edges}{Edge list as a matrix or data frame with two or three columns. The first two columns
represent the vertex indices of each edge. Edges must be represented twice, once for each
direction. The optional third column represents edge weights.}

\item{...}{Additional arguments passed to the \code{UndirectedGraph} constructor for setting up the
C++ mdgm::RNG. Can set the \code{seed} as an integer.}
}
\value{
An \code{UndirectedGraph} object.
}
\description{
Creates an undirected graph from an edge list
}
