% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdgm-package.R
\docType{package}
\name{mdgm}
\alias{mdgm-package}
\alias{mdgm}
\title{mdgm: Mixture of Directed Graphical Models}
\description{
Provides a C++ implementation of a mixture of directed graphical models with R bindings.
}
\author{
\strong{Maintainer}: Brandon Carter \email{brandon.carter@duke.edu}

}
\keyword{internal}
